Contributed initial Xilinx Spartran II components, all fitting in a 208 pin PQFP.
These symbols/components breaks a number of rules, but they also show the
weakness of these rules for components of this type.

Powerfeeds is explicitly shown since these chips actually have a total of 9
diffrent powerlines and 8 diffrent voltage references. The intended
funtionallity of the IO-pads changes with the powerline hookup so it becomes
part of the design.

The symbols/components where made in a big square with pinning spread as they
are on physical components. This aids proofreading and even more important,
actual usage of schematics in a lab setup.

Making and proofreading these symbols where not an easy effort, so errors are
bound to exist. However, all symbols have been proofread to verify correct
pinout. These symbols helps to show us real difficulties and what we need to
make smoother.

Ref:
Xilinx DS001-4 (v2.3) March 5, 2001 "Spartran-II 2.5V FPGA Family: Pinout
  Tables".