Allow gschemdoc to access wiki pages with -w . Applied patch from Peter Brett. * scripts/gschemdoc.sh: Apply patch from Peter Brett to allow gschemdoc to access wiki pages with -w .
The original patch initially tries to find a local export
of the wiki; on failure, it falls back to the online wiki
at the gEDA website.

Since the online wiki and the user's geda version may not
be the same, I commented out the part to display the online
wiki if the file is not locally available. Thanks!