Implement new component library system in libgeda.
A new component library model is needed to allow gschem etc.
to work well with part database systems.  This patch
provides the necessary infrastructure in libgeda, without
the frontend Guile commands required to make use of the new
features.  In particular, it abstracts the component
library into the concepts of "Component Sources" and
"Symbols", and allows external "Library Commands" to
provide symbol data in addition to the traditional directory
libraries.

In order to allow the new system to work, there are major
changes in all code which manipulates COMPLEX objects; in
particular, the magic "EMBEDDED" string is now only used
in the on-disc representation.