Make comp. library search faster and more flexible
Replace s_clib_glob() with s_clib_search(), which can operate in
either exact matching or glob matching mode.  Use a hashtable to cache
the results of library searches. The hashtable is cleared every time
it is invalidated (for instance if a component source is added or
removed, or the component sources are rescanned).