Changed/updated the #define mechanism for creating custom gEDA/gaf versions.
Previously this mechanism was used to create custom versions of gEDA/gaf.
With the coming of stable and development versions of gEDA/gaf, this
mechanism was changed so that end users can tell the difference between
a stable release vs a development release (by looking at the version string).
The CUSTOM_VERSION #define was changed PREPEND_VERSION_STRING and is now
prepended to any strings that output the VERSION string.

Changed PREPEND_VERSION_STRING to 1.0- since this is the stable-1.0 branch.
Had to resolve a cherry-pick conflict in gattrib's about dialog box since
it changed a bit more on the master branch.