Portability fixes for the testsuite.
Get the testsuite fully working on SunOS-5.9/sparc.  This
entails two issues.  The "-I" option to diff is a GNU
extension and is not supported by the solaris diff.  Work
around this by using sed to filter out the lines which were
to be ignored.  This addresses the tests run by gnetlist/tests/runtest.sh.

The other issue is that the longopts tests in the spice-sdb
testsuite should not be run on a system that does not have
getopt_long().  To address this, extend the run_tests.sh to allow
a variable in the tests.list file to disable a test unless set to "yes".
(cherry picked from commit f8680b18c8e27617eccbbf158acb84ccac7ac386)