gschem: Fix a race condition with mouse wheel scroll events (1740327)
If more than one mouse wheel scroll events are received into the event
queue before the first one can be processed, then the program was
zooming in to the wrong location.  This is because the mouse moved
event handler, which is triggered by a call to warp the cursor in the
scroll event handler, must run to update the mouse postion global
variables that the scroll event handler needs.