Implemented the ability to change the color of junctions (code for #1746675).
The filled circles (at the junction of nets) and filled boxes (at the
end of unconnected nets) were the same color and it was hard to tell the
difference from a distance.  This patch addes (junction-color ...) rc
keyword and allows the user to change the junction color.  The default
color for the dark color scheme is yellow and the default color for the
light color scheme is purple.