Fixed drawing bug when using attribute autoplacing.
When placing a new component, if attribute autoplacing was enabled,
the refdes was drawn wrongly, and it was shown fixed in the screen despite
the component being moved.
This was due to attribute autoplacing function was called before drawing
the component. This function erases the old text (which wasn't drawn), thus
drawing the text in the screen permanently.
This fix disables all drawing functions when calling autoplacing function,
and renables them afterwards.