Extend GschemDialog to allow hooking to save/restore internal geometry
GschemDialog now emits "geometry-save" and "geometry-restore" signals
which instances can connect to. Subclasses may over-ride and chain the
"geometry_save" and "geometry_restore" class members. These signal handlers
are passed a pointer to the GKeyFile and a text string identifying the
group it expects data to be saved under.

The code relies on features only available in GLIB 2.6 onwards. When using
this functionality the code must be surrounded with an appropriate test as
the class functions are not defined when compiling against an older GLIB.