Fix zoom extents to work correctly first time. Closes bug #1699959.
A page's visible area is intialised to be the same size as the canvas
extents. The zoom extents code scales based on this visible area, however
this initialisation isn't correct, since the aspect ratio of the screen
doesn't always match the canvas extents. The first call to zoom extents
would fix this by correctly setting the visible area.

By making the zoom extents code work from the screen aspect ratio (scaled
to world coordinates), we can zoom to the visible extents in one shot.

Also fixed some /2 operations, which for doubles should be /2.0