Make the preview widget resizable
Added handler for the configure event to resize the backing store and
re-zoom the preview. Added the A_PAN_IGNORE_BORDERS flag to the zoom
extents calls - giving a nice centered preview.

Altered component selector dialog's alignment widget to allow the preview
widget to expand to fill the available space.