Alter panning code to center the page in the viewport if under-zoomed
Centering is independant in the X and Y directions. When the viewport is
larger than the page canvas in a particular direction the page is
re-centered in that direction. This is necessary for a nice preview.
Also change preview widget code to remove the A_PAN_IGNORE_BORDERS flag
from zoom operations, which is no-longer needed with the above changes.