Change color in the multi-attrib dialog "value" entry when insensitive.
Makes the "value" GtkTextView look the same as its neighbouring widgets
when insensitive. Unfortunately GtkTextView always appears to render using
the style's GTK_STATE_NORMAL text color. We save that initial value, and
copy across either the saved value, or the GTK_STATE_INSENSITIVE text color
when setting the widget sensitivities.

NB: This is just a workaround for what appears to be a bug in GTK+.