Fix for Bug#1776547: graphical versus source
The usage described in Bug#1776547 is pretty unique, but it might be handy
to support in general.  The recursive source= and graphical= behavior is
a unexpected side effect of the traversal of graphical symbols/components
(552e06e921e4b6eb758b9483d1913c39b3fef5f5).  The fix is to explicitly
check for the graphical= attribute and if it is found to not further
traverse the hierarchy.  Regression tests were added so this behavior
doesn't change in the future.