Moved refresh button in component select dialog box
The refresh button was along the very bottom of the component select
dialog box at the same level as Close, Apply, and OK.  The refresh
operation is not nearly as important or first class as these actions.
The refresh button was moved to the filter area.  Whether this change
sticks depends on what users think of the new location (I'm quite willing
to move it elsewhere if there is a better spot).