Fix for Bug# 1782032: Blank attibute name when attribute value starts with " "
In order to fix this bug, we needed to validate the input attribute when:
1) input an attribute using the single attribute editor,
2) Add an attribute in the multi-attribute editor
3) Modify the name or value of an attribute in the multi-attribute editor
An attribute is not valid if the name or value are empty, name ends with a
space, and/or value starts with a space.  o_attrib_get_name_value does
this validation throughout gaf.
(cherry picked from commit b018b88a1f2a88d246af850a67a15be5b19825b2)