Fix memory corruption and possible segfault copying attributes
If an object (but not all its attributes) was selected and copied, dangling
pointers were left in the un-copied attributes. A subsequent copy operation
containing those missed attributes would attempt to reattach them to the
copy of their parent. If this had been deleted in the mean time, memory
corruption and a possible segfault would occur.

Consolidates the duplicated code which sets the ATTRIB property "copied_to"
into a helper function, o_attrib_list_copied_to(). This is also used to
clean up after copy operations.