Fixed a whole bunch of gtk+ 2.4.x and misc issues building on an older box
This is a fix for (Patch#) 1789290 libgeda-1.2.0 fprintf which supplied a
similar patch which fixes the issues in libgeda/src/s_clib.c and
libgeda/src/s_textbuffer.c.  In the future, all releases will be built with
gtk+ 2.4.x so that these sorts of problems do not get released.