Fix for dialogs drifting downwards on Cygwin/X.
On Cygwin/X, reopening a gschem dialog restores its original position,
but with a small vertical offset.

The workaround is to show the dialog before restoring its position.

The docs for gtk_window_move() offer a possible justification:
  "Most window managers ignore requests for initial window positions
  (instead using a user-defined placement algorithm) and honor requests
  after the window has already been shown."

The downside is, if you have a slow system, you could see the dialog
appear briefly at the old position, I suppose.