Fix a bug in o_attrib_search_name_single() were it might return NULL.
If o_attrib_search_name_single() doesn't find an attribute attached to
the passed OBJECT, it checks if the passed OBJECT is itsself an attribute.
In this case, it would return NULL (the result from the preceeding search).

The bug might not actually be triggered in normal usage, as most callers
don't pass a variable to recieve the OBJECT pointer.