Remove the ATTRIB linked list in favour of a GList
The ATTRIB.copied_to pointer is moved from the ATTRIB struct to reside in
the source OBJECT being copied.

The ATTRIB *attached_to pointer of an attribute's text OBJECT is changed
to OBJECT *attached_to, pointing directly at the OBJECT to which that
attribute is attached. This avoids the need to traverse the ATTRIB list
and retrieve this pointer from the HEAD attribute node. (A HEAD node is
not used in the GList of ATTRIBs).

The ATTRIB structure now only contains one element, OBJECT *object. This
points to the text OBJECT holding and displaying that attribute's
"name=value".