Update the all drawing state SCREEN co-ordinates when zooming / panning.
By updating all of the co-ordinate state when zooming / panning, a redraw
which includes a draw of any rubberbanded items will show them in the
correct location. This allows XOR drawing to continue without dropping
artifacts, or flashing up a "ghost" where the original element being moved
was. The explicit erase for the gost item when moving is no longer needed.