Fix Bug #1838164. Initialise copied_to pointer in objects to NULL.
Without explicitly initialising OBJECTs to have a copied_to pointer of
NULL, copying a selection containing an attribute but not it's parent
object could cause a segfault. The code checks to see where the parent
object has been copied for correct attachemnt of the new attrib. If
the parent object wasn't copied, this pointer should be NULL.