gschem: Fix new page behaviour
Currently, gschem attempts to load "untitled-1.sch" from the cwd at
startup.  This is clearly wrong -- gschem should present a blank page
regardless of whether there is a page available on disk (consider
behaviour for "untitled-N.sch", N>1).

This has the side effect of stopping gschem from trying to load backup
files for new pages (this is a good thing).