clib: Speed up caching by using pointer rather than string keys.
The initial version of the symbol cache used the symbol name as a
string as the key, mostly because of the obnoxious 'const' keywords I
had put in the definition of s_clib_symbol_get_data().  Having found a
way around that, using a pointer to the symbol structure as the key is
much faster as it avoids costly string comparisons and allocation of
heap memory (as well as being slightly more elegant).