Fix test of object_changing in o_grips_motion() to a g_return_if_fail()
This test isn't hit unless o_grips_motion() is called before the
object_changing variable is set, and this would indicate a bug in
libgeda or its client. This could currently cause a segfault if it
were to happen, as the variable isn't initialised to NULL.

Initialise object_changing to NULL, so any bug will always be caught.

Tidy up indentation levels in o_grips_start() by returning early if we
don't have any object to manipulate, rather than wrapping most of the
funciton in an if (...) {...} test.