Always redraw the screen using a blit from the backbuffer.
The previous redraw scheme which would draw to the screen concurrently
with drawing to the backbuffer is proving to be a bottle-neck to recent
desktops using compositing. Every draw we make forces the X server to
compute a damaged region and re-composite the window.

We now draw only into the backbuffer, and then invalidate the extents of
the region we've drawn.