Change image embedding format to base64 encode the original files
Previously, the format was a base64 encoding of serialised output from
the gdk_pixbuf used to store the image internally. Due to another bug
(now fixed) it is known that no official gEDA release could re-load
the embedded images that it saved. This hopefully provides some
justification as to why its OK to change the embedding format here.