Changed libgeda's configure mechanism to be similar to gschem's configure.ac.in
In order to allow a wide range of gettext versions to be used, gschem has
a configure hack that dynamically figures out the gettext version and creates
configure.ac from configure.ac.in.  This hack is now in libgeda with the
additional improvement that if configure.ac.in changes, configure.ac and
configure are automatically recreated.  Added a special mode to autogen.sh
(recreate_configure_only) to support this improvement.