Added checks at the autogen.sh stage for missing m4 macros
This commit adds a autoconf friendly m4 macro to check for undefined
macros (originally written by Peter Brett).  This macro is used in
configure.ac.in to check for a few troublesome macros (guile, nls,
gettext, intltool) that are usually missing because a required -dev
package is not installed/missing.