De-select list in the compselect dialog when placement is cancelled.
By de-selecting the list, we ensure that the list state reflects the
component the user may currently place. It allows the same component
to be selected for again, without either using the "Apply" button,
or having to temporarily select a different component first.

As a hook to de-select the list is added in o_redraw_cleanstates(), we
now avoid using that call as a shortcut to cleanup our last place
operation. It is still used if the last operation was not a component
placement.