Added the use of PCBDATA environment variable to make gsch2pcb relocatable
Before this patch, m4_pcbdir was being set from the preprocessor variable
PCBDATADIR which was set at ./configure time.  In order to be able to move
gsch2pcb to a different install prefix, this path has to be changable at
run-time.  If PCBDATA is not set, then PCBDATADIR is used (previous
behavior).