Fix spice-sdb slotting to work without modified pinseq attributes
Removes all previous code which iterated over slots for a given package,
replacing it with a custom (get-uref ...) procedure which ensures each
slot of a component is given a unique uref when gnetlist traverses the
schematic. This means each slot will appear as a separate "package" in
the gnetlist data-structures, even if each slot uses the same refdes=
attribute.

The format for the package uref is "(refdes/uref).(slot)", in keeping
with the existing spice-sdb slotting code.