gnetlist: Strip trailing lower case suffixes from PCB package names
Adds a custom (get-uref ...) function to the PCB, pcbpins and gsch2pcb
back-ends. These strip lowercase suffixes from the uref, to ensure that
slotted package names like IC1a and IC1b are treated as the same, IC1
by gsch2pcb.

This suffix stripping matches the fact that PCB ignores the lower-case
suffix on net names. It would have been possible to leave the suffixes
in the PCB and pcbpins output, although in the PCB netlist case, this
causes any implicit power nets in each slot to be duplicated in the
netlist, which PCB reports as an error.