Removed misleading "pin count from footprint=" attribute check
This check wasn't working correctly for all footprint= values.
It probably worked fine for the DIP# and related footprints (ending
with the number of pins in the footprin), but for others it was giving
incorrect info/warnings.  This check will be replaced with something a
little bit more flexible going forward in a later release.