update the pinnumbers if a slot attribute is deleted
Update the slots whenever a text line that is attached to an object
gets deleted. The update checks the slotting even if there's no slot
attribut. This catches the case when the user deletes a slot attribute
and the current displayed pinnumbers are not the ones defined in the
symbol. This commit fixes [#1880221]