Invalidate component cache when updating symbols. [#1880211]
Temporary fix for Bug #1880211, "symbol update reads the symbol from
the cache". Ensure that when refreshing symbols, the symbol is read
from the component source, not from the cache. (Suggested by Peter
Brett).

A more complete fix will be to add API which invalidates a single symbol
name, rather than the entire cache.