Only call gdk_window_invalidate_rect() when drawing to a GdkWindow.
For image export, a GdkPixmap is used in place of w_current->window,
and calling gdk_window_invalidate_rect() with this as an argument
causes lots of nasty warnings. It led to a crash when print.scm is
used to export from the command line.

This "fix" is really quite a nasty hack. Since w_current->window is
a GObject in both cases, a test with GDK_IS_WINDOW() can be used to
distinguish the cases. A better fix will come from future rework of
the drawing code.