fix drawing artefact when drawing a net and zoom/pan
Zooming and paning redraw everything and if there are rubbernets
they get deleted, too. We have to tell the rubbernet function that
the rubbernet is no longer on the screen.
Thus I've added a flag rubbernet_visible as indicator. This flag needs
to be set when drawing, and reset when the rubbernet gets erased.