connect the magnetic net mode into the net drawing functions
The magnetic net mode is now use to set the start net point, track the
closest connection in the rubbernet function and to get the end point
of the net. The endpoint in the magnetic mode is calculated in the new
function o_net_finishmagnetic.

This commit is based on the patch written by Franz Mottas and Thomas
Arndt [#1824420].