Add a common set of tests to gnetlist evolved from the spice-sdb tests.
The tests are applied to all backends, with control files being used to
document the expected return codes, and golden files with the expected
output. These files are generated from a 1.5.0 development release of
of the gEDA suite, and do not in all cases represent "correct" behaviour,
merely "gEDA 1.5.0" behaviour.