use tiles to find connections for magnetic net mode
This commit changes the magnetic net code to use the tiles.  This
reduces the load when checking all the possible connections for every
mouse movement. The patch adds a library function to get object lists
which may contain objects in the region of the magnetic reach. In
o_net.c the object iterator code has been changed to use the new library
function.