Make the pads netlist backend use CR/LF instead of just LF.
Based on a discussion on the gEDA mailing lists it seems that PADS
(which runs only on windows) needs DOS style netlists and not unix
style netlists.  So rather than forcing a unix2dos step in beween
netlisting and using the result, go ahead and teach the pads netlist
backend to directly write out a DOS style netlist.  Update the
testsuite accordingly.