guess net direction when adding nets in gschem
When drawing a L-shape net the old behavier was to draw the horizontal
net first and the vertikal after. This patch guesses the direction
dependant of the starting point of the net. It tries to continue nets
and not to overwrite them.  This functionality was originally written
by "ar" and posted to the geda-dev mailing list.