Use o_complex_rubbercomplex_xor() where possible to simplify code.
Adds a parameter "int drawing" to this call, which is passed on to
o_drawbounding(). This is used to determine whether the operation
is to undraw an old object (in which case, the old feedback mode is
used), or is drawing new objects, in which case the current action
feedback mode is used.