tragesym: fixed version string handling
With the commit 5a35c238f19622f8045784fa61497c80c4144fa6 a trailing " 1"
was added to the version string. This is only correct if the source files
do not contain that part of the version definition.
Changed tragesym to accept both version strings "20060113 1" and
"20060113" in the source file. Bug pointed out by Stefan Salewski.