gschem: use pointer from drawing area in preview and x_attribedit
The preview widget used in the file dialog and in the component dialog is
now using the pointer from the drawing area, too. The motion event function that tracked the pointer is no longer required.

The attribedit dialog used the mouse_wx/y variables to set the attribute
postion when activated with the hotkey. This usage has been removed.

Removed all mouse_wx/y occurences in x_event and in the global.h file.
This variables are completly gone now.