libgeda: fixed EOL handling for mixed windows/linux usage
A file that is created on windows with windows EOL characters
may not be read into gschem without errors on Linux. This fix
strips off leading and trailing CR characters when reading a
single line from the textbuffer.
Bug reported by Thomas Arndt [#1993952]
The bug was introduced with commit 40833189228f8edeff