Improve the search for already opened page in s_hierarchy_down_schematic_single(). [#1909932]
Fix for Bug #1909932, "gschem: page manager confused about hierarchy".

When looking for a sub sheet in s_hierarchy_down_schematic_single(),
the filename may not have been suitable for comparison to opened page
filenames. It then failed to identify a sub sheet as already opened
and opened it again. Adds normalization to filename prior to
comparison.