Altered point selection mechanism to use distance instead of just a bounding box.
This patch improves the mechanism used to select objects when using a single
point.  In addition to using the bounding box, this patch calculates the
distance between the selection point and the object itself.  This calculated
distance provides a more accurate means to determine the actual object the user
selected.

Signed-off-by: Peter TB Brett <peter@peter-b.co.uk>