gschem: Activating component selector rows expands / contracts groups
Observation of a user working with gschem showed that opening the groups
of components from varous libraries was not particularly discoverable to
a user not familiar with GTK. The user was noticed to be double-clicking
the library sub-groups, rather than identifying the expander arrow.

Fix this usibility issue by making double click, and other row-activation
methods expand / contract the library groups. Keyboard navigation
(where space / return activate given rows) is also easier now.