Make the "gain" of zoom in / out operations an adjustable parameter.
The new gschemrc function, (zoom-gain integer) controls the percentage
size increase when zooming into the page. Zooming out uses the inverse
factor such that a zoom in followed by a zoom out will return the
schematic to the same size.

E.g:
   20% increment => x 1.2 original size when zooming in
                 => x 1 / 1.2 x original size when zooming out

The default gain keeps gschem's previous hard-coded behaviour of zooming in
by a factor of 1.5. This is now specified with the command (zoom-gain 50).