gschem: Move freeing stretch list of to after move operation is ended
Previously, the stretch list was left stale with pointers after a move
operation, and only cleared when commencing a new move. By clearing it
after each move operation (completed, or cancelled), we can ensure its
contents are always valid.

This is important so that at the end / cancellation of a move operation,
we can operate on the stretch list without danger it was not up-to-date.
If a move operation is started with rubberbanding disabled, and the user
subsequently activates rubberbanding (during the move) with a key-stroke,
stale pointers in the stretch list may have been dereferenced in error.