gschem: Only add nets and buses to the stretch list.
Only nets and buses are rubberbanded, so there is no point adding other
connected items (such as pins) into the stretch list.

This fixes a bug where pins attached to nets being moved disappear if
the user pans whilst moving. (The objects in the stretch list are marked
"don't redraw" when starting a move operation.)

To avoid nesting (yet another) huge "if" clause in the length of the loop
in o_move_check_endpoint(), the loop has been re-written as a "for" loop
to ensure its iterator is incremented when the loop is "continue"'d.
The various nested : if (test) {...}          // wrapping the loop's core
are now:             if (!test) continue;