libgeda: Add field to TEXT for storing displayed string.
Sometimes, the string to be displayed is different from the whole
string contained by a TEXT object (for instance, when an attribute has
either its name or value unshown). When using "real" fonts we don't
want to fix this up each and every time the string is rendered, so
it needs to be stashed somewhere.