gschem: Pass dialog to multiattrib treeview data funcs.
The cell data funcs for the multi-attribute editor dialog's treeview
need to have access to a libgeda TOPLEVEL to be able to use
o_text_get_string(). The obvious thing to do is to pass in a pointer
to the dialog (which contains a pointer to the GSCHEM_TOPLEVEL, etc)
as the user data pointer.