o_complex_add: Remove un-needed save/restore of PAGE->object_tail (CODE!)
(* FOR GOODNESS SAKE PETER, TRY TO COMMIT THE CODE WITH THE COMMIT!)

toplevel->page_current->object_tail was saved and restored around
symbol data loading and subsequent attribute promotion.

The PAGE->object_tail is not altered during this block of code.
This was experimentally verified during significant "real work"
usage conducted with a g_assert() prior to the restore step.