libgeda: Add bounds validity flag.
Adds a flag to OBJECT marking whether they are valid (i.e. whether
they've been updated since the last operation which modified it). Each
of the per-type recalc functions is updated to set the flag, although
nothing clears it yet (this will be in a later patch).

When world_get_single_object_bounds() is called, it checks the flag,
and if not set it calls the appropriate recalc function. If the flag
is -still- not set, it fails.