gnetlist: Don't drop all hierarchy prefix for children without refdes=...
If a hierarchy block is missing its refdes=... attribute, insert it
without any change of hierarchy prefix. The block's contents keep the
same hierarchy level as the instantiating schematic. Previously the block
was inserted at the root level, with no hierarchical name mangling.

This may be useful in some cases if you want to suspend hiearchy mangling
for certain portions of a hierarchical design. (NB: You would then be
responsible for avoiding refdes name clashes). A similar effect can be be
obtained on a design wide basis by setting the options:
(hierarchy-*-mangle "disabled") in a gnetlistrc file.