Revert "name = value" space handling in o_attrib_get_name_value()
commit 872287f23b4880b01babed777efe7976783feab2 re-wrote this function
using UTF-8 string manipulation functions, and changed a test for " "
characters either side of the attribute's "=" to use g_unichar_isspace().

This g_unichar_isspace() also tests positive for tabs and newline
characters, making previously legal attributes illegal after this commit.
Revert usage of g_unichar_isspace() in favour of single byte tests.

An important property of UTF8 encoding is that the 7-bit ASCII codes are
not used within bytes of a multi-byte character. Given the start of a
chacacter in UTF8, testing if the first byte compares to the 7-bit ASCII
code we want is a sufficient test.