libgeda: Tidy implementation of o_attrib_attach()
Various cleanup and restructuring of the code. Should be identical
in functionality, but with an extra g_critical warning instead of
a silent return in the case attrib is not found in parent_list.

Changed printf to the console for g_warning or g_critical messages.
Some of these cases where we can't attach an attribute probably ought
to be caught in gschem, in which case we could make these warnings a
higher severity, such as g_critical.