Revert commit #95923f7de815ec58819eaf94ab9332d51a28fbb9
This commit, "Simplify prototype of s_tile_remove_object()."
removed code to make a page being deleted the current page.

This change introduced a crash when closing gschem with multiple
pages open. We shouldn't really need a page to be current to delete
it's objects, but this is an underlying bug which should be fixed
before this part of the patch is re-applied.