Move calls to attribute promotion code from inside o_complex_add()
Previously o_complex_add() would perform attribute promotion or delete /
hide eligible attributes depending on the "attribute_promotion" argument
passed to o_complex_add(). Break these into separate operations:

1. Get / add to OBJECT list, a single complex object with o_complex_add().
   This object has its prim_objs exactly as the symbol exists on disk.
2. EITHER:
     Promote the eligible attribtues with o_complex_promote_attribs().
     (For GList based processing use o_complex_get_promotable(), attach
      the returned objects as attribs and place in the appropriate GList.)
   OR:
     Delete / hide eligible attributes with
     o_complex_remove_promotable_attribs()

All behaviour relating to attribute promotion should be identical after
this commit. If not, it is a bug should be reported. (This includes the
insertion order of the OBJECT and it's attributes into the object list).