gschem: Don't add invalid attributes to the multiattrib dialog list
Since we can't edit them, don't treat them as attributes.

Gschem validates attributes upon creation, so an invalid attribute
attached to an object must come from a hand-edited file, or a text
object which has been attached to an object. (Validation that the
text item is a valid attribute isn't performed in that code-path).

In the future we might want to add validation of attributes at
file-load and at before attaching text OBJECTs to a COMPLEX.