Fix f_print_set_line_width() to emit valid postscript
Previously this function emitted "%d mils setlinewidth\n". The "mils"
portion seems to be left over from previous printing code, as "mils"
isn't defined to be anything in prolog.ps, nor in the headers we emit.
Our units are scaled to mils, so we can just emit "%d setlinewidth\n".