Move some static variables inside o_grips.c to GSCHEM_TOPLEVEL
This is necessary for grip manipulation of more complex OBJECT types
such as the forthcoming PATHs. As the entire drawing state can't of a
path can't conveniently be stashed in the GSCHEM_TOPLEVEL variables
such as first_w{x,y}, second_w{x,y} etc., we need to reference the
OBJECT and grip being manipulated from the drawing functions which
are outside o_grips.c.