Added code to do something more sane when trying to print tabs
Previously if a tab character was encountered in a text string, the
postscript output code would output a Y with a dot on top.  This is
clearly not a tab character.  The patch adds explict handling of the tab
character to output 8 spaces instead.  Not perfect, but better behavior
than before.