gschem: code and test file for the net selection patch
This commit adds the main functionality to gschem. The new function
o_select_connected_nets() will select nets depending on the configured
net_selection_mode and the current net_selection_state.

The implementation uses two stacks for traversing the connections. The
netnames stack contains all connected netnames and the net stack
contains a collection of all currently collected nets.

Note: The search for the netnames is very expensive, because of the
required search of all netnames in all objects.
The default net_selection_mode in system-gschemrc is set to "enable-net",
only the directly connected nets are selected.

The functionality can be testet with the test schematic
tests/net_selection.sch.

This patches are based on the work of Franz Mottas [1895538].