gschem: show attributes in the component select dialog [#1444029]
gui code to show the attributes in the component select dialog. The
attributes are taken from the TOPLEVEL structure of the preview
widget.

The attribute treeview is only created if the GList of the attributs
is not empty and the attributes are only updated if the treeview is
present.

Added an option that will print all attributes in alphabetical
order. To do that the first element of the option has to be and "*":
(component-dialog-attributes '(*))