Get rid of s_color_get_index().
This function is very clearly the Wrong Thing, and could potentially
result in incorrect color indices ending up in schematic files if two
or more color map entries have the same value.

This patch removes the function entirely, and modifies the Scheme code
which uses it to simply pass color indices around instead. Note that
the results of s_color_get_index() were never actually *used* in the
existing code.