Revert applying -Werror to compiler flags when linking with GCC (CODE!!)
Depending on the gcc version, the original developers might not see the
warning until the software gets into the end user's hands (and tries
building it with an older/newer version of gcc).  Most warnings are
not catastropic, but by using -Werror they are _all_ catastropic.

If a developer wants to get rid of all warnings, they should either:

 1) run make -s and see/fix all the warnings, and/or
 2) modify the Makefile's themselves and add the -Werror flag temporarily.

(Comments explaining rationale snipped out of an email from Ales).