gattrib, gschlas: Restore setting dummy select_func to ensure proper save
The select_func function pointer (now owned and exported by libgeda) is
assigned to OBJECT->sel_func for non-locked objects. Locking is achieved
by setting OBJECT->sel_func to NULL, and when writing out objects, the
locked flag is determined by comparing that that pointer to NULL.

We need to keep providing a dummy function so we can differentiate
between that and a NULL pointer. This was accidentally lost in
commit 1aab94354f053887db67aa5702abc595795dfe82 which moved the
select_func global variable into libgeda.