libgeda: Attempt to fix f_nomalize_filename for win32 systems
Patrick's re-written code for this made the assumption that the
root portion of an absolute path in windows is 3 characters long,
e.g. C:\. This is an invalid assumption to make.

Unfortunatey, g_path_is_absolute() regards paths such as \foo as
absolute, from which our code then assumed it could strip off the
first 3 characters. There are also UNC paths such as \\server\share\
where the root portion is an arbitrary length, which would not be
handled correctly.

g_path_skip_root() is aware of these details, and could be used to
pick out the root portion of a path accurately, but it doesn't help
in all cases. We could use it to determine that we need to add a
current drive-letter to a "semi-absolute" path without one, however
there is also the case with a path such as "C:foo", where we would
need to insert the CWD from a particular drive.

Path            ->  Root
----                ----
\\srv\share\foo ->  \\serv\bar\  (fine, path is absolute)
C:\foo          ->  C:\          (fine, path is absolute)
C:foo           ->  C: (?)       (needs current directory from drive C)
\foo            ->  \            (needs current working drive or UNC share)
foo             ->               (needs current working drive and directory)

Rather than trying to implement the logic to normalise paths on
Windows, conditionally compile in usage of the Win32 API function
GetFullPathName() when _WIN32 is defined. This implementation is
slightly modified from the Windows portion of libiberty's lrealpath
function. (GPL V2+).

f_normalize_filename() now differs slightly in its sematics on Win32,
as GetFullPathName() doesn't require or check that the file exists
on disk. We could add this feature if it turns out to be required.